HADOOP-7610. Fix for hadoop debian package. Contributed by Eric Yang
git-svn-id: https://svn.apache.org/repos/asf/hadoop/common/branches/branch-0.20-security@1167379 13f79535-47bb-0310-9956-ffa450edef68

+++ src/packages/update-hadoop-env.sh
@@ -133,7 +133,9 @@ if [ "${UNINSTALL}" -eq "1" ]; then
     rm -rf ${PREFIX}/etc/hadoop
   fi
   rm -f /etc/default/hadoop-env.sh
-  rm -f /etc/profile.d/hadoop-env.sh
+  if [ -d /etc/profile.d ]; then
+    rm -f /etc/profile.d/hadoop-env.sh
+  fi
 else
   # Create symlinks
   if [ "${CONF_DIR}" != "${PREFIX}/etc/hadoop" ]; then
@@ -141,7 +143,9 @@ else
     ln -sf ${CONF_DIR} ${PREFIX}/etc/hadoop
   fi
   ln -sf ${CONF_DIR}/hadoop-env.sh /etc/default/hadoop-env.sh
-  ln -sf ${CONF_DIR}/hadoop-env.sh /etc/profile.d/hadoop-env.sh
+  if [ -d /etc/profile.d ]; then
+    ln -sf ${CONF_DIR}/hadoop-env.sh /etc/profile.d/hadoop-env.sh
+  fi
 
   mkdir -p ${LOG_DIR}
   chown root:hadoop ${LOG_DIR}
